+++ packages/jest-worker/src/__tests__/index.test.ts
@@ -5,6 +5,7 @@
  * LICENSE file in the root directory of this source tree.
  */
 
+import {pathToFileURL} from 'url';
 import type {JestWorkerFarm, Worker, WorkerFarmOptions} from '../';
 import type FarmClass from '../Farm';
 import type WorkerPoolClass from '../WorkerPool';
@@ -69,6 +70,19 @@ it('makes a non-existing relative worker throw', () => {
     // eslint-disable-next-line no-new
     new WorkerFarm('./relative/worker-module.js');
   }).toThrow("'workerPath' must be absolute");
+});
+
+it('supports URLs', () => {
+  const workerPathUrl = pathToFileURL(__filename);
+
+  // eslint-disable-next-line no-new
+  new WorkerFarm(workerPathUrl, {exposedMethods: ['foo', 'bar']});
+  // eslint-disable-next-line no-new
+  new WorkerFarm(workerPathUrl.href, {exposedMethods: ['foo', 'bar']});
+
+  expect(WorkerPool).toHaveBeenCalledTimes(2);
+  expect(WorkerPool).toHaveBeenNthCalledWith(1, __filename, expect.anything());
+  expect(WorkerPool).toHaveBeenNthCalledWith(2, __filename, expect.anything());
 });
 
 it('exposes the right API using default working', () => {